demo/requirements.txt
flask
flask-cors
flask-login
flask-mail
flask-sqlalchemy
flask-wtf
geoip2
jinja2
lxml
matplotlib
maxminddb
numpy
pandas
pillow
pyshark
python-docx
python-dotenv
requests
rich
schedule
seaborn
selenium
sqlalchemy
tqdm
werkzeug
wtforms
